+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## To Be Released
+
+### Added
+- Support plain string in first arg. `orn 720p 1080p`
+
 ## [2.0.0] - 2016-09-27
 
 ### Changed